move components out of systems mod into a componensts mod
