+++ src/camera.rs
