+++ src/components/mod.rs
@@ -1,1 +1,3 @@
 pub mod path;
+pub mod ball;
+pub mod videographer;
