+++ src/systems/ball.rs
@@ -1,48 +1,12 @@
 use amethyst::{
-    assets::Handle,
     core::Transform,
     derive::SystemDesc,
-    ecs::{Component, DenseVecStorage, Join, Read, ReadStorage, System, SystemData, WriteStorage},
+    ecs::{Join, ReadStorage, Read, System, SystemData, WriteStorage},
     input::{InputHandler, StringBindings},
-    prelude::*,
-    renderer::{SpriteRender, SpriteSheet},
 };
 
 use crate::config::WanderballConfig;
-
-pub struct Ball {
-    pub radius: f32,
-}
-
-impl Component for Ball {
-    type Storage = DenseVecStorage<Self>;
-}
-
-pub fn initialize_ball(world: &mut World, sprite_sheet_handle: Handle<SpriteSheet>) {
-    let mut local_transform = Transform::default();
-
-    let (ball_radius, view_height, view_width) = {
-        let config = &world.read_resource::<WanderballConfig>();
-        (config.ball_radius, config.view_height, config.view_width)
-    };
-
-    local_transform.set_translation_xyz(
-        view_width - (view_width * 0.25),
-        view_height - (view_height * 0.75),
-        1.0,
-    );
-
-    let sprite_render = SpriteRender::new(sprite_sheet_handle, 0);
-
-    world
-        .create_entity()
-        .with(sprite_render)
-        .with(Ball {
-            radius: ball_radius,
-        })
-        .with(local_transform)
-        .build();
-}
+use crate::components::ball::Ball;
 
 #[derive(SystemDesc)]
 pub struct BallSystem;
@@ -60,7 +24,7 @@ impl<'s> System<'s> for BallSystem {
         let movement_y = input.axis_value("move_y");
         let fast_movement = input.action_is_down("fast_movement");
 
-        for (transform, _ball) in (&mut transforms, &balls).join() {
+        for (transform, _) in (&mut transforms, &balls).join() {
             if let Some(mv_amount) = movement_x {
                 let scaled_amount;
                 if let Some(true) = fast_movement {
