+++ src/systems/mod.rs
@@ -4,6 +4,5 @@ pub use self::path::PathTileSystem;
 pub use self::videographer::VideographerSystem;
 
 pub mod ball;
-pub mod camera;
 pub mod path;
 pub mod videographer;
