+++ src/systems/videographer.rs
@@ -2,54 +2,18 @@ use amethyst::{
     core::transform::Transform,
     derive::SystemDesc,
     ecs::{
-        Component, DenseVecStorage, Entity, Join, Read, ReadStorage, System, SystemData, World,
+        Join, Read, ReadStorage, System, SystemData,
         WriteStorage,
     },
     input::{InputHandler, StringBindings},
-    prelude::*,
     renderer::Camera,
 };
 
 use crate::config::WanderballConfig;
 use crate::side::Side;
-use crate::systems::ball::Ball;
+use crate::components::ball::Ball;
 use crate::util::{point_near_edge_of_rect, Point};
-
-/// The entity that holds the camera and moves it when it needs to
-#[derive(Default)]
-pub struct Videographer {
-    pub view_width: f32,
-    pub view_height: f32,
-    pub view_x: f32,
-    pub view_y: f32,
-}
-
-impl Component for Videographer {
-    type Storage = DenseVecStorage<Self>;
-}
-
-pub fn initialize_videographer(world: &mut World) -> Entity {
-    let (view_height, view_width) = {
-        let config = &world.read_resource::<WanderballConfig>();
-        (config.view_height, config.view_width)
-    };
-
-    let videographer = Videographer {
-        view_height: view_height,
-        view_width: view_width,
-        view_x: view_width * 0.5,
-        view_y: view_height * 0.5,
-    };
-
-    let mut local_transform = Transform::default();
-    local_transform.set_translation_xyz(videographer.view_x, videographer.view_y, 2.0);
-
-    world
-        .create_entity()
-        .with(videographer)
-        .with(local_transform)
-        .build()
-}
+use crate::components::videographer::Videographer;
 
 #[derive(SystemDesc, Default)]
 pub struct VideographerSystem;
